Update HashString function to ensure hash fits within int32 range
This commit modifies the HashString function in the hash.go file to use int32 for the hash variable, ensuring that the computed hash value fits within the int32 range. This change improves type safety and prevents potential overflow issues when hashing strings.

+++ pkg/utils/hash.go
@@ -1,10 +1,10 @@
 package utils
 
-// HashString creates a simple hash of a string
+// HashString creates a simple hash of a string that fits in int32 range
 func HashString(s string) int {
-	hash := 0
+	hash := int32(0)
 	for _, c := range s {
-		hash = 31*hash + int(c)
+		hash = 31*hash + int32(c)
 	}
-	return hash
+	return int(hash)
 }
